[roll] Roll fuchsia [memory] Split lazy computation of Inspect output
Split the LazyNode computation of memory_monitor Inspect output in
two parts, one for the fast, already-stored data, and one for the data
that need a full capture first. This ensures that we can output as much
data as we can in Inspect even if we hit a timeout.

BUG=365100782

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169832
Original-Revision: 38ea2372689969031e9784181f862edaa7082a07
GitOrigin-RevId: 0c17a2bee00bfd41b44153b537dd127138e22040
Change-Id: I26a4eb73a01c31b02a5332296d9def66501563a7