[roll] Roll fuchsia [debugger] Attach to jobs at ComponentStart
Attaching to jobs during ComponentStarting notifications is critical to
attach as high as possible in the job tree for a particular job_only
filter. If we only attach during process starting events, we will only
ever attach to the immediate parent job of that process. The desired
outcome for a job_only filter is to attach as high in the job tree as
possible to stay out of the way of any descendant processes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157212
Original-Revision: 0ebebaff9f1f0f4b48325c9d63fddda924cd8da7
GitOrigin-RevId: 65c00c08acc1722d400aff25901cc4bffe0a77d1
Change-Id: I146f12abe096d7d8bb1738abd9e32509f4a16231