[roll] Roll fuchsia [netstack2] Replace connect with r*
We hope to simplify the io2 rights model in the future by removing
explicit connect rights. We also want to make sure that service and
diagnostics directories can be enumerated, and also traversed. To
achieve this, we need to replace the use of "connect" rights in all CML
files with "r*".

These rights are *not* required for connecting to actual service
instances, however using a consistent set of r* rights allows various
client libraries (e.g. fdio) to assume a basic set of directory
operations on a given service directory. Right now this is not possible
as fdio does not use *any* rights, so the current set of rights are not
enforced to begin with.

Original-Bug: 346585458
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170096
Original-Revision: e72df01a542ec7f1953a05e85002af66495039e8
GitOrigin-RevId: d4e66cfd81044055d2a977ba0ad40ae49d441fc4
Change-Id: I0694c85b36eaee7d7c8f8befd7457cc56369c58a