[roll] Roll fuchsia [component_manager] compose a dictionary of dir_entry for service caps
When a service capability is routed, return a dictionary instead of a
dir entry. This dictionary is populated and kept up to date using a
filesystem watcher on the service directory in the capability provider's
outgoing directory. This will allow future commits to perform instance
renaming, filtering, and aggregation on dictionaries.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136135
Original-Revision: e8fe578d890bdd3ab2be889b1043319870ed81b4
GitOrigin-RevId: e2a8299df5db156f0787b1edb2a7f66080e33333
Change-Id: I92d1f7987eba40c18df32861e0862f37322f7862