[roll] Roll fuchsia [ffx][emu] Derive emulator mac address from name if possible
When the emulator is started with `--name fuchsia-5254-X-Y`, and X and Y
allow to construct a valid mac address, it will not be generated by
hashing but instead immediately converted from the name. This is helpful
for OTA testing with the `--uefi` flag by ensuring that the target
nodename does not change across reboots.

Original-Fixed: b/383061309
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170873
Original-Revision: 8c779b432bc7e991170cab2f04d8dcec7b072eee
GitOrigin-RevId: a7f538179740e7e690481580f884d49f49e50e7c
Change-Id: I535c2135e0558a782b4371fb13033febc8fbbfe7