[roll] Roll fuchsia [fuchsia.sys2] Rename RealmQuery.Open -> DeprecatedOpen
Although RealmQuery.Open was removed at API level 25, there still remain
a lot of usages throughout the codebase, and new instances still keep
cropping up. In an effort to discourage this and to make identifying
the remaining instances easier to find, we rename Open to
DeprecatedOpen.

Likewise, we follow the same pattern for the deprecated StorageAdmin
method OpenComponentStorageById. To improve readability of the FIDL
files, we also move the deprecated/removed protocol functionality to the
end of the protocols they are defined in.

Although this requires rewriting history, this shouldn't be a breaking
change for any clients, since the method was already removed at API
level 25, and all servers are in the platform.

Original-Bug: 331322369
Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162332
Original-Revision: 6c1ebdc930c87f582cd130e8c8804deaf2434175
GitOrigin-RevId: 3392c0be14d473b3cb8569248abd04de451f6de7
Change-Id: I7ad88658afea19ba1efd68c4a9ad7a49724b733f