[roll] Roll fuchsia [memory] Optimize memory_monitor's `GetCapture` allocations
On devices which use the `StarnixCaptureStrategy`, there is a lot of
known memory overhead due to Starnix sharing VMOs between Linux
processes; the `Strategy` itself (which gets re-processed on every
`Digest` call, e.g. when calling `ffx profile memory`) grows quite a
lot.

This change uses the fact that an instance of `CaptureStrategy` in
practice never needs to be reused; this lets us `std::move` its
tracking state instead of `std::copying` it when consuming it, saving
the related memory.

This change saves ~15MiB of allocations on at least one device that is a
heavy user of `Starnix` processes, when profiling `memory_monitor`
during a call to `ffx profile memory`. It should both be safe and save
some CPU as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168339
Original-Revision: e411de87281a673b9009c133e7e951c5e432685e
GitOrigin-RevId: c186dde769cfeab7c71a2160df13d8bf8ac5f867
Change-Id: Ia1c8445e456ed1ab6029c55556c4eb1a14117db9