[roll] Roll fuchsia [ffx][product] Improving product download error messages
This cleans up the error messages when the product bundle
cannot be found to download. This also fixes a problem
when the product bundle could be duplicated in the list.
Now the list is de-duped before reporting an error.

Original-Bug: 383123090
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171244
Original-Revision: 6ff01b9f148a623fcc5760438bbaa77c1340a7b2
GitOrigin-RevId: d7560697dd7391611cb66d5d6e4bf6e35fef57af
Change-Id: Ia56b4596ba67dc40dc176ec942ecb75ae94af4b6