[roll] Roll fuchsia [ulib][c] Begin soft migration of files to /strings
https://github.com/llvm/llvm-project/pull/118899 began the process of
moving files from /string to /strings. This CL likewise introduces the
/strings directory to Fuchsia's build machinery and performs the first
step in migrating the needed targets to /strings.

Since it's expected that we'll need to repeat this process as more
targets are migrated to strings/ in the future, this CL introduces a
(temporary) exec_script that can be used as an example for future CL
migrations, to make it easier and faster to detect file changes (but
note the exec_script gets cleaned up after each migration, as is done
in Ia8de544a2695125f888544071b08f37906fe381d).

Original-Bug: 383347626
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172248
Original-Revision: 575cd6a3d6a2829f5c91f5a8d1cf199d42ea8808
GitOrigin-RevId: a2a59a5095415d8eacdfb6a1a8c4268102471e4d
Change-Id: Ie7d1fbc863db617cac3d7119994b0f9ba7103545