[roll] Roll fuchsia [sestarnix] Turn todo_check_permission() from macro to function
Encapsulate the details of a "todo_deny" check into a structure so
that the `todo_check_permission` API can be turned into a function.

The structure is still created by a macro, primarily to validate
the bug URL format, but it's a simpler one, but decoupled from the
actual check API, for greater flexibility.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171774
Original-Revision: 2721234182a2524cde7211da161184d1feabb26a
GitOrigin-RevId: 79b25dc93b8c62355ac862e0ac21ad504347a3ad
Change-Id: I25b46d9818940c91c6054881fbe4936833d599fe