[roll] Roll fuchsia [sestarnix] Introduce has_file_permission() helper
has_file_permission() accepts a FileObject and a set of file-like
permissions. It first validates that the current task has the
fd:use permission to the FileObject, before delegating to an
has_fs_node_permission() helper, which ensures that they have the
requested permissions (which may be empty) to the underlying file
node.

Original-Bug: 364569337
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169485
Original-Revision: 123020a5f83e4ce68a884aa19b779292467456f8
GitOrigin-RevId: ea594f6e84e02162d7e72fad1b686a8e057c50a3
Change-Id: I6599c80b69d04f840dfc96abd62b918e3068e7d0