[roll] Roll fuchsia [bazel_sdk] fix analysis failure in fuchsia_api_level
This CL adds the ability to use the flag that we define for pigweed
which skips fuchsia targets to opt out of the analysis failures that
occur when we have an unset api level. This happens because in pigweed
they run builds with `bazel build //...` which ends up running the
implementation of the fuchsia_api_level rule to figure out what the
build config value should be. We cannot avoid running this implementation
so we use the flag they already have set to skip validation.

Original-bug: 354047162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173506
Original-Revision: 1747704cf2b00fcf64c29e63cf3a59a00601d091
GitOrigin-RevId: f64cb901b8fa41a8a9eabcfbf5862697cd3fccd0
Change-Id: I44207a7ab870edb3271195d1ab17b417dc280160