[roll] Roll fuchsia [kernel][vm] Simplify LruThread reasoning
Tightens the coordination between the lru_event_ being signaled and the
LruThread performing processing to simplify the reasoning. The previous
logic of checking for the lru signal without the lock and performing
one processing attempt per wakeup should be correct, but attempting to
reason through a proof of this is difficult and so is worth simplyfing
the code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171874
Original-Revision: 5c3e46fc80b249cd52b92133333f95fcb025bbdf
GitOrigin-RevId: 65ba7ee28f7cbec2a4a57aa51b94c3457fa06b7d
Change-Id: I17b92ccf68b0cc70f3f042dc61d874cc0c6209f6