[roll] Roll fuchsia [sdk][ffx][starnix] remove Suspend() and Resume() FIDLs
Now that Starnix Runner is fully responsible for suspending
and resuming Starnix Kernel, we no longer need to provide
FIDLs for other components to do so.

Moreover, continuing to provide those FIDLs is a liability,
as calls to those FIDLs may conflict with the autonomous
logic in Starnix Runner.

Given this, we remove the obsolete and potentially problematic
FIDLs.

Because these FIDLs were never published at a stable SDK
version, they can be removed from the source entirely (versus
adding a `removed` annotation).

Removal of the `ffx` tooling is a little more delicate. But
we were unable to find any source code that depends on it.

Original-Bug: 378470841
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170856
Original-Revision: 8d84f4b8c5b333449ebe4df6a3f35af7a78094fe
GitOrigin-RevId: fe7464471572763548baf9c10481ae5e563fa7bf
Change-Id: Ief67f861917a95b88d87745a0d360434cba0d70c