[roll] Roll fuchsia [driver][power] Acquire lease to honor timeout
Make sure the system stays awake until we reach the timeout asked
for by the client, but be lazy about acquiring a wake lease. We
may obtain a wake lease when we are informed the system is
suspending instead of when the client calls `HandleInterrupt`.

Previously, if the system was resumed we didn't obtain a wake
lease. If the system started suspension later, before the
requested timeout, we would not stop it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173498
Original-Revision: 85a9bb7b6bb83c18430c032cf73f28b4e31cf7b1
GitOrigin-RevId: 2f768b46f9ff5f0569bbbb602d6581f50665b4a9
Change-Id: I5f6e1c67009a4c42d220d2fc5d5b669a8c0a3841