[roll] Roll fuchsia [fint] Ensure that `fint set` calls `build/regenerator`.
A reland of [1] after fixing the un-handled command-line flag that
caused the revert (+ associated regression test)

---- original commit message ----

This is necessary before adding any more logic to the regenerator
script. For example, [2] tries to move the generation of the Bazel
workspace to build/regenerator, but fails with an error when invoking
Bazel that says that the workspace does not exist.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1174452
[2] https://fuchsia-review.googlesource.com/c/fuchsia/+/1169075/3

Original-Bug: 382019521
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175036
Original-Revision: 8ec8644a5db480d8a872e00bd31859a04430f25a
GitOrigin-RevId: fcc8372256b700c2674ec64d2a405fcb4a1461d3
Change-Id: Idb6303b8a191b24787952a8af78aba623bcb0c6e