[roll] Roll fuchsia [build] Add //build/testing/python_build_time_tests.gni
Add a new GN template to define a target to run a python
test at build time. This should be only used to verify the
correctness of low-level scripts used by the build, where
early failure detection is critical.

This is a generalization of the //build/api:run_all_tests
target that can be used to verify other scripts.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175632
Original-Revision: 7bd090810b5888509dac525b4fcb49c194176681
GitOrigin-RevId: 36d04c40d9cb180b010db36ca489f5934208ee95
Change-Id: I3d5fd7f0099a25860a1000deec114798fcc81e9e