[roll] Roll infra/recipes [fint] Add `compilation_mode` to replace `optimize`
"optimize" was originally chosen as the field name for configuring debug
vs. release builds based on the assumption that the `optimize` GN arg
would eventually be used for specifying debug vs. release but that
hasn't happened and `optimize` is used for a different purpose. So
rename the `optimize` fint parameter to `compilation_mode` to avoid
confusion with the `optimize` GN arg.

A soft transition is required to avoid breaking the infra recipes and
configs, so keep accepting `optimize` for now.

Original-Bug: 382773442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170792
GitOrigin-RevId: 26f844ce2125920d7180be45024c7d4ad2bca048
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1176235
Original-Revision: dbd4a06be219979108f3ab2f68c83f60a62c9f2c
Change-Id: I0f9215803cbea6444324d169cee0790de11f3806