[roll] Roll fuchsia [test_manager] enable debug_agent backtrace
This CL turns on test_managers integration of debug_agent, which causes
a backtrace to be sent to stderr when a test panics. This backtrace
is displayed in fx test output and is available in test output
artifacts.

Currently, when a rust test fails, we see two stack traces, because the
rust runtime also dumps the backtrace.

Test: feature is covered by current unit tests

Original-Fixed: 286555768
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152601
Original-Revision: f397837ac8e3b4c4990e85c16be5821749a3e749
GitOrigin-RevId: 862244ad78a90c609877a655be3923ea7c0bdb7b
Change-Id: I8135e39a5e1226db8f2d36d6e3bba6e19fa81d90