[roll] Roll fuchsia [kernel] Count mappings at phys hand-off
This change updates the arch VM backends to count the present mappings
made within the kernel address space mappings handed off from physboot
to the PMM. Currently the kernel's own boot-time MMU code accounts
for this, and so this is needed for parity (since that code is going
away).

This allows us to remove a switchback in the x86 boot MMU code to also
handle the counting of handed-off tables. This simplifies that code as
well as the related start.S logic to save the physical addresses of
some handed-off tables.

We further implement this for for riscv64, for parity and simplicity's
sake, even though wider mapping accounting is not yet in place.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174073
Original-Revision: 4763d8158e76aab135a561a6f50500cb4a27b5f7
GitOrigin-RevId: 59be9615ab3384d771c817ad8f3687b32495793f
Change-Id: I46ec20653c7f8c78c81e00205c08fdaab33950a3