[roll] Roll fuchsia [starnix] Reset the deadline in the heap after the timer is expired
`current_deadline` is checked when a new timer is added to the top
of the heap. When it's the same as the existing one, no driver action
will be taken. Android might add the same timer again after the timer
is expired. Therefore,the `current_deadline` should be cleared every
time when the timer is expired.

Original-Bug: b/384103431
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175073
Original-Revision: 2dd64f268983aa9ac565b7285a1857514370491e
GitOrigin-RevId: 0d6b690db19ad2957d5e6bcf42659809bbba6034
Change-Id: Id33621b9bb2bde678f5183cdae48cd12dcd1e6db