[roll] Roll fuchsia [virtio-gpu-display] Use display::ModeId and display::ModeAndId.
This CL updates the C++ adapter interface to reflect the expected usage
of display modes in the Banjo/FIDL interfaces between the Display
Coordinator and engine drivers.

The implementations hard-code a ModeId of 1, or expect a ModeId of 1.
This matches the expected implementations of engine drivers that support
a single preferred mode.

Original-Bug: 384573759
Test: fx test //src/graphics/display/drivers/virtio-gpu-display
Test: Booted workbench_eng.x64 on qemu
Test: ffx target ssh display-tool squares
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1176095
Original-Revision: 80887a12e4dfa964519ea4c4f6003cc74a19df6a
GitOrigin-RevId: 767a32cfb819409bf4d0059e5d5d41bd4e1f4915
Change-Id: I9c5eab76a527274de532880cdbefe758a413a5ff