[roll] Roll fuchsia [audio][drivers][aml-g12] Calculate VMO size from requested format
`GetVmo` calculates the required VMO size from the requested
minimum number of frames and the ring buffer format.

Previously, it used the format set in `engine_config.ring_buffer`
that is set in `ResetEngine` from the DAI format.

This happens to work if the client first sets a DAI format that
matches the requested ring buffer format passed to `CreateRingBuffer`,
but calculates the wrong size in two cases:

1. Client calls GetVmo before setting DAI format
2. Client uses a ring buffer that is supported, but does not match
   the previously set (or default) DAI format

Case #1 is useful to support because it allows the client to power
the SoC on using `SetActiveChannels` before setting the DAI format
so the call takes effect.

ADR calls `GetVmo` once the ring buffer is created. Without this
change, ADR rejects the VMO because the VMO size is too small.

Original-Bug: b/384110895
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175802
Original-Revision: a844f3a6ea2ea253c7fa2bf93beb9023317a248b
GitOrigin-RevId: 934504d3a0df9b81ceaea41609f8e644ecee9177
Change-Id: I982e98f1d0e51470d29b746b410078d28e1cff41