[roll] Roll fuchsia [dl] libdl can run initializers
Introduce `RuntimeDynamicLinker::RunInitializers` that takes the module
returned by a successful `Open` result and runs initializers on that
module and its dependencies. Initializers are run in reverse order of
the dependency graph (i.e. the reverse load order) so that init funcs
for dependencies are run before the dependent module.

This does not include support for synchronization or handle re-entrency,
which will be addressed in a future change.

Original-Fixed: 382340693
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174713
Original-Revision: 8bbd767c142bf19d819d9b632fae7447a24ebcd2
GitOrigin-RevId: 53a7e124b6a4cd8f69082e183f2c9659d9e769ab
Change-Id: I13f7b0dbd89505eac47bf949e6b7aedffa7e5742