[roll] Roll fuchsia [ffx][emu] Assign host ports when using qemu
This adds assigning the host port to a port that is forwarded out of the
guest. Older versions of QEMU automatically assigned an unused port to
complete the mappings, however QEMU 9 does not do this and so
ffx does this as part staging the instance.

Original-Bug: 384957072
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178082
Original-Revision: a48b6fb36bfaddca5289ddc0c4a33cccf584397f
GitOrigin-RevId: f805794e268b2036befec5f3ff07edc8db7fbd2e
Change-Id: Iaa2dcd95b4c010c6f1c98ac3fe610d25f7f2f0f6