[roll] Roll fuchsia [wlan][storage] add golden file test
Add a golden file unit test to check that WLAN can load a file with a
particular format. The test is meant to remain the same even the format
for the file changes, as long as the old version is supported.

test: fx test wlan-storage-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177093
Original-Revision: 87dfb5f582a3fcbc48cd89ac29c839ad9be0e10c
GitOrigin-RevId: d01e2807cb2c34206b406bf601e2f1479c2d693d
Change-Id: I13ac0053118cfddf5197726bba7ff3014ba38876