[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/lib/fuchsia-sync/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180363
Original-Revision: 770b71125f838dee37fdbc185c6c20e9ff911a36
GitOrigin-RevId: 097333bbd8185d8763accc4fc6146b3f1902c604
Change-Id: If5aea27576839755755f0dd39d086ab03776211a