[roll] Roll fuchsia [owners] Remove anmittal@google.com from src/tests/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180352
Original-Revision: 5429f13a72635b4930ff4a9f1883f52bdaaee56d
GitOrigin-RevId: 330fc24c92183ce3691fe450ee48969e19456b1c
Change-Id: I58a0a94307420ea1915205825cace4a3692da694