[roll] Roll fuchsia [owners] Remove anmittal@google.com from tools/testing/testparser/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180371
Original-Revision: fd41085685d3c04cba78547285a5987754dd849f
GitOrigin-RevId: f2cb233dd487fa3e2d7eed5422be5c6747924407
Change-Id: I2ac839f722549f780fa90e2ac4b1af6a4853dad0