[roll] Roll fuchsia [owners] Remove tmandry@google.com from sdk/rust/zx/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180365
Original-Revision: 87e3258cc411ad1cc3952a5a0d1a32752e8aad67
GitOrigin-RevId: 09a7acf6806c78a8c0c03021756d4a1617854fa8
Change-Id: I5e9f6b6918345a6a543f535411b5e02f88b7da7f