[roll] Roll fuchsia [owners] Remove tmandry@google.com from docs/development/build/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180360
Original-Revision: adf6201d37fea35f08089c2577bb511eb706b451
GitOrigin-RevId: 92b15e0c28f63d3a4c2ed579e1a6d84df19abdfb
Change-Id: I27f1dabacb4c1d079fbb00fc6c575e2a900e55c9