[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/lib/fuchsia-async/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180357
Original-Revision: 32da3b668c5e51c1a78dd7f05dcb347be9e777d5
GitOrigin-RevId: 376a0258ebd94e76e0b77311940179ab924fe2b5
Change-Id: Ied612fb1a2254d16bfeae860017780cbbffed477