[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/lib/cstr/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180372
Original-Revision: c8aebeb20229528b30cde994e706c78db32b693c
GitOrigin-RevId: fc157e80e2c82ebdc4a3fa7242bb9c0c2acb7131
Change-Id: Id2a0cc65a1f2a91fc40610b5f56d68c625121078