[roll] Roll fuchsia [ffx][doctor] Removing dep on ffx_command and global Env
This removes the dependency on ffx_command, which was used to
get the target spec. It was replaced with calling
ffx_target_get_target_spec.

Since this call needs an EnvironmentContext, this was added to
the FfxTool struct, and the global context usages replaced.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181013
Original-Revision: 07ec4722453342a8c850418ef32e01917ac046c0
GitOrigin-RevId: 159ba6da5f1a7acf4d99a49429e1b90aa41682ac
Change-Id: I7388c142c76069ff609fe1858796889b75d8f94e