[roll] Roll fuchsia [kernel][phys] Construct mappings in physboot for ELF kernels
This change begins constructing and handing off the kernel's virtual
mappings for ELF-loaded kernels. We are already doing this in the legacy
codepath for x86, and so need to do this now to follow through on the
transition to ELF-loaded x86 kernels. Although in any case we'll have to
do this eventually anyway and it's easy enough now to do!

This means that for all architectures on which we hand off the kernel
page tables and load the kernel as an ELF - which at the moment is only
riscv64 - the kernel has its final protections from the start. In
contrast, the old code made those all-permissions mappings initially
and only changed them later in VM startup. The later VM post hoc
proctection logic is still in effect, but it's just a no-op in this case
now.

Moreover, we start dynamically linking the x86 kernel in the ELF
codepath at its intended virtual address (0xffff'ffff'0000'000), which
begins to get exercise (along with the above functionality) in the
physboot-elf-kernel boot test.

Note that we're still branching to the ELF kernel at its identity-mapped
address.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173134
Original-Revision: e11e43a0aec639408cd0e284228517faf3008b81
GitOrigin-RevId: 140d1cfd3afedc1439bd18dfa529b4bf092c6be8
Change-Id: I15f523bd37fa0cb9456cb727b9f93d3672449b56