[roll] Roll fuchsia [starnix] CloneHelper only inspects errno on MAP_FAILED.
The assertion that errno == 0 has been causing occasional flakes in
the waitpid_flags test suite. However it's not guaranteed that
errno will be equal to 0 when mmap succeeds, so we should not assert
on its value. It seems likely that the flake is caused by a previous
test leaving a non-zero errno value before this code executes.

If the mmap is actually failing, then we will still detect that by
checking for MAP_FAILED and logging errno's value if that's the case.

For now, mark the flake as fixed and let it get re-filed if the mmap
is actually failing to cause the flake.

Original-Fixed: 387167196
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181772
Original-Revision: 62521d98b7389572a7fa84702276c23af8247160
GitOrigin-RevId: fd6b684c726a417d4ebad6f97d3a5e4b9f3776f3
Change-Id: Ia2f862507f17d6eea57d34ff505b865abd431a78