[roll] Roll fuchsia [fuchsia-component] Avoid opening the instance directory
Rather than open the service instance directory, we can construct an
opener which is capable of opening the service instance member relative
to the service directory. This is more efficient. This change does
require that we wrap the Service object in an Arc as the service proxy
takes a `Box<dyn MemberOpener>` which requires a static lifetime.

This change also changes fidl service proxies to provide a method for
getting the instance name they represent by extending the MemberOpener
trait.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155213
Original-Revision: 56b3ae5af800e2deac6c71491f41673953e281bb
GitOrigin-RevId: cb89b77418dd45870516594dbe247c5e12f0e2b0
Change-Id: I74a0d11daa12049515949dafd3fc869a67f9a89e