[roll] Roll fuchsia [owners] Remove tmandry@google.com from tools/devshell/contrib/lib/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180353
Original-Revision: ddecacca7e6b6426f775cb350e621328bf75060f
GitOrigin-RevId: 032dba304edc141bf4b1fb840548cec07ea01fec
Change-Id: I8d228dee97e01cd101e29c9481690646fd872a24