[roll] Roll fuchsia [kernel][arm64] Reuse sp in arm64_elX_to_el1()
arm64_elX_to_el1() is updated to install the current stack in SP_EL1
before dropping to EL1 so that it may be reused. This should be a no-op
today, but will enable normal use in C++ once this routine is repurposed
for use in physboot.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182356
Original-Revision: 089375d32b1be7c6d9d8324b8973c036b81de583
GitOrigin-RevId: 183a0e2bb0985b399c1770f02f0dd3304cd90f65
Change-Id: Ib5f96f8858627c53ddca084db1873d97f4241e3c