[roll] Roll fuchsia [owners] Remove tmandry@google.com from tools/shush/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180362
Original-Revision: 0fc2aa417979f7bdec96d0e77cda8fef0489c615
GitOrigin-RevId: 83333746736699f0989c79fc9fad702eaf75a51b
Change-Id: Ife81c083ebaa81892dd2c992c73542fb70717158