[roll] Roll fuchsia [owners] Remove anmittal@google.com from zircon/system/ulib/svc/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180359
Original-Revision: 95136509ffea2f7410e7084297e99e8a0fc0af99
GitOrigin-RevId: 65414054ece801ed133435ec478a9b4843c100cf
Change-Id: Id8000e8b2097ea715ba0300a28506dc97824bcda