[roll] Roll fuchsia [bazel] make compdb point to fidling directory
When bazel collects the arguments for c++ files that take a dependency
on a fidl library it uses a path which includes a virtual includes
directory. This directory does not point to a valid location in the
fuchsia tree. For now, we can just point to the GN created fidling
directory so clangd picks it up.

A future change should look into why the virtual includes does not exist
so we can avoid manipulating the arguments more.

Fixes: 382260446
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182362
Original-Revision: 145ae8cdab96498616f1168f49e1000c1c385dbc
GitOrigin-RevId: e64cc98b46ca19f2d8ed8da7a131a7fd7e1dfc23
Change-Id: Id19e129f388a1d2666d2e82846e4965c95be3648