[roll] Roll fuchsia [owners] Remove maheshsr@google.com from tools/devshell/python/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182614
Original-Revision: 185162c465ee5f42a7ef660e4d08762eeb71fc9a
GitOrigin-RevId: ec924d9c8bdc69536a221e11fbdbffcf6e0db57d
Change-Id: I105130c2467e72e5d46f4924007c3a5c369240a2