[roll] Roll fuchsia [owners] Remove maheshsr@google.com from scripts/lib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182613
Original-Revision: 360a99f62ae82f80b03ca8d6b834251580db310d
GitOrigin-RevId: 57296358c544b83abd3947a464c81383834acc08
Change-Id: I7104f93a542b44ab5788af45f90f751e2266cd86