[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/isort/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182597
Original-Revision: 15a62f7060ed0570420fc131df37f7b7ff591a36
GitOrigin-RevId: ca0dd00cbee3463a5b8fa19c29f51e88ddb3f364
Change-Id: I7d379be5aaa51925d01db372cea2eccff90f1809