[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/mypy_extensions/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182598
Original-Revision: 47d4f92d52f788a77b5e50d8c4fe10745338aabe
GitOrigin-RevId: 8a2712a7529216accf772b5c9d3b94e3a7a9a75b
Change-Id: Ia6deb7bd00dddf188a319402165987dd896f2f01