[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/.cargo/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180366
Original-Revision: 54e9299d072eb7edffd3c0d6b606184db3465ab4
GitOrigin-RevId: 17cecde06b06fa23adad0a615db580351ac013f1
Change-Id: I2d89be7c301a0830becbe5b2e779c86dfa9f1c86