[roll] Roll infra/recipes [testing] Run triage tool within the triage sources root directory.
In a normal build, when we download the orchestration_inputs, we store
the triage sources as absolute paths. However, when using local build
artifacts, we store the triage sources as relative paths. Thus we need
to run the triage tool in the triage sources root directory so that the
relative paths work.

Original-Bug: 372309107
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1182062
Original-Revision: ea4fdc28f708cc6181859da08d081d9450199a2b
GitOrigin-RevId: 6f6fdd6813a79fdf3150eb57299fe3174367b9c0
Change-Id: Ie5e4d0d5c27877ea24858a27834bc1d7738fcb85