[roll] Roll fuchsia [kernel][arm64] Disable EL2 monitor and virtualization tests
Doing so is the plan of record, enabling a more straightforward rollout
for physboot to load the kernel at its ultimate virtual address at EL1.
In particular, this in effect disables our arm64 kernel hypervisor (to
eventually be re-enabled more simply to rely on ARM's EL2 host feature).

This is a simpler reland of fxrev.dev/1176695, which was reverted as a
sequence of changes that seemed to introduce secondary boot failures on
a post-submit builder. More testing was done with this piece to screen
for such failures.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182332
Original-Revision: 84c93cd4c7adde7e2a88f05f1289c3ba71c45861
GitOrigin-RevId: 05c0f04a655c98949cf8b0929154c8edbd2ddccd
Change-Id: Iaeef8375663a92294f6dbb12394c6efa6b25cb3d