[roll] Roll fuchsia [kernel][phys][arm64] Hand off address space(s)
As already done with riscv64 and x86, this is a major step toward
physboot properly creating the kernel's address space: it hands off the
address space as it is currently constructed, upper and lower - along
with the kernel's mapping with its final page protections in place -
enabling the kernel to bootstrap from the already-constructed identity
map.

This is a reland of fxrev.dev/1173133, which was reverted due to
instability around secondary boots. It is now believed that this was
due to stores of boot info on the boot CPU not being flushed in time for
loads on secondaries for consultation. This new version addresses that.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182458
Original-Revision: 87a3a481b6abe928e978c680a378b86c8e98c64d
GitOrigin-RevId: 5d775f83123c5679b4a192f88452c55b801f4a1a
Change-Id: Ibdb00083cf5c5a9fb639a57ed039c9ca7708144c