[roll] Roll fuchsia [owners] Remove tmandry@google.com from third_party/rust_crates/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180367
Original-Revision: 22d5ef063fb988918436eeb071fef69c5a6fc2a9
GitOrigin-RevId: 2f1caddcc3cea074ac84cb53a11106f0bbc02613
Change-Id: Ibc40bc1a7c8301e88510b3ef0c0ee49a537177c0