[roll] Roll fuchsia [owners] Remove tmandry@google.com from docs/development/languages/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180368
Original-Revision: 2f53b4c4414dda3c184bf0268df1d92052c888d3
GitOrigin-RevId: 17d5fa92552f961bfeea23fc6036da037bec0c21
Change-Id: If848f01dd8053a7e2ed657e8bb838cce41cc17a1