[roll] Roll fuchsia [fdf][power] Use dynamic offers for power protos
This works around the limitation described in 369189827. We cannot use
dictionaries to provide protocol offers that exist in the static offers
of a cml. This makes the protocols that we want to override into dynamic
offers. This allows us to control whether we offer these from dynamic
offers here, or through a dictionary in the next cl.

Original-Bug: 369189827
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164958
Original-Revision: 194e64c0bc14291cc93cc3de3a119271b4e46632
GitOrigin-RevId: f1e21f265b9780484565186101f819d09b55b95d
Change-Id: I586dba20e637d37ef1ae37e8148466baf014b12c