[roll] Roll fuchsia [fxfs] add any range prefix to allocate list immediately
In allocate, if there is a section of the range we need to allocate
before the extent we found, do that right away so all codepaths after
get this case handled for them. Before this, existing overwrite extents
would cause allocate to skip this section of the range.

This also leads to a big simplification to the extent switching range
math that happens later.

The unit tests are expanded to make sure all the allocated ranges
actually end up with overwrite extents covering them, which confirms
there are no other versions of this bug.

Original-Bug: 293943124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1184618
Original-Revision: 9a885eb4e61a865e43e50e714672afffbdf7ec09
GitOrigin-RevId: 8ec6e162400609941c1c8a3cf462ae4b961b4ae2
Change-Id: Id285b4cc1de69d76599308cce582cce8690aa745