[roll] Roll fuchsia [starnix] Allow SIOCSIFNETMASK with no addresses
SIOCSIFNETMASK is strictly supposed to return an error if invoked with
no addresses installed on the interface, but a current workaround to
support an upper-layer product workflow needs this to succeed instead.
Once blackhole interfaces are supported, this can be removed.

Original-Bug: 385171649
Original-Bug: 387998791
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185473
Original-Revision: 789a9fb6d818137c77fb981809be0a495340230d
GitOrigin-RevId: 2de3312d671e5b12080d5a98151ff934483a28fe
Change-Id: I1c824678cda678400d74358805aa046c8c0f7f80