[roll] Roll fuchsia [eBPF] Type-safe eBPF arguments
Previously it was possible to call `EbpfProgram::run()` with an
arguments of arbitrary types, which may be different from the types
used when the program was verified.  Now it ensures that the passed
argument types match the types that were provided when the program
was verified. Expected argument types are defined in the new
`EbpfProgramContext` trait that replaced `EbpfRunContext`. The
argument types must implement `ProgramArgument` trait which allows
to map Rust types to eBPF types. In most cases types are checked
when a program is linked in `link_program`, except for conformance
tests where the size is not known in compile time. Conformance
tests use `link_program_dynamic` which checks argument type when
`run()` is invoked.

Also `PacketAccessor` trait has been replaced with `Packet` trait
that has to be implemented by the packet type. The interpreter
uses `FromBpfValue` trait to convert `BpfValue` to a packet. This
allows to avoid unsafe type casts outside of the eBPF intepreter.

Original-Bug: 385009220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177632
Original-Revision: b04f4cf0f5dd24556bde43948684622fdf9a4f20
GitOrigin-RevId: dbf54cc25bf7e448dcd8a93b549787a3407b3119
Change-Id: I0dbad039f1146d0147b5841e370240a8550dd05c