[roll] Roll infra/recipes [run_postsubmit_tryjobs] Don't let launched builders outlive parent.
If the run-postsubmit-tryjobs builder is canceled, we should cancel the
child builds as well. For example, if using the Mega-CQ button which
launches this builder, uploading a new patchset will cancel the build
but the launched subbuilds will continue to run. They should be canceled
along with the parent build. See for example go/bbid/8726354788034027233
where the build got canceled, but the launched subbuilders kept going to
completion.

Test: go/bbid/8726171464340188449

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1186039
Original-Revision: 7558d7cf5cb91217022f39a048a57c0ce505eb1c
GitOrigin-RevId: 0219f5ab4e6e9546b0bf136bd25b3915913d368e
Change-Id: I39869855fc67a649cff437948f3c48afa33dc29b