[roll] Roll fuchsia [kernel][vm] SetStreamSize unmaps
Calling SetStreamSize unmaps the range from min(new_stream_size,
old_stream_size) to the end of the VMO under the object lock.

This is required to avoid a race when a VMO is mapped with
ZX_VMO_FAULT_BEYOND_STREAM_SIZE and the stream size is reduced. Without
unmapping the pages, it is possible to sneak in a read before the page
is zeroed with ZeroRange.

Run-All-Tests: true
Original-Bug: 379565554
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169435
Original-Revision: 83b35d17a8053bfe20a4f32caa08c6feee55b1dc
GitOrigin-RevId: 945f7e154a232cd70a8f9f5267a8268448985b01
Change-Id: I108e89210281a5b8701120eb15e14639f0df30a0