[roll] Roll fuchsia [audio] ActivityReporter for new usage
Convert protocol ActivityReporter from closed to open, so
as to add a method that clients can use to watch for
activity on AudioRenderUsage2.

In this CL, to minimize churn to clients, the new method is
strict and only render activity is updated. A future CL will
convert the method to flexible and add a companion method
for capture streams.

This is all in service of the new AudioRenderUsage2 enum
value, ACCESSIBILITY.

Original-Bug: b/384972490
Original-Bug: b/384932846
Original-Bug: b/371061877

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180653
Original-Revision: ed4e77c7dc969caeeccab0fed4560d0c0ba3b256
GitOrigin-RevId: f244e86029f4ce4e4786a32f0c05b60f8d1487f2
Change-Id: I4fe9b6f37fc1e54c99f0c8ac2f3c71e47fa5dfb0