[roll] Roll fuchsia [cml] Update duplicate offer message to reflect type of offer
When we have a duplicate offer to "all" with a child component,
the error message should reflect that it's a child component
and not a capability. This previously caused confusion
during a migration where there was both a dictionary
called "diagnostics" as well as a component called "diagnostics".

Original-Fixed: 383624158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185612
Original-Revision: 1dca1cf073ae0aaf96d110789222434329e6cc55
GitOrigin-RevId: a16928aa2f63521797b57104764734bb63bffd39
Change-Id: Ibadd562d4062b8465f972ab0bb9594e071ad7d54