[roll] Roll fuchsia [assembly][diagnostics] Fix serialization issue with assembly overrides
The serialization format didn't match the original format and was
leading to an error when overrides were used.

The reason this was done is that we can't send objects in structured
config, so the object was being serialized into a string. This assumed
that we never did deserialize(serialize(deserialize(S))), since before
this change serialize(deserialize(S)) != S. However, we do that when
dealing with assembly overrides.

Instead this change guarantees that serialize(deserialize(S)) == S and provides a function to create the string used for the component'
structured config.

The associated bug has the exact repro scenario.

NO_IFTTT=no changes to the the config schema

Original-Fixed: 389747855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187292
Original-Revision: 817cedfb074374a2be766a23dfaee97d81f16074
GitOrigin-RevId: f09901e5f59fe5c9f6405e303b5723e5e1c3e5e5
Change-Id: Id8b5056fc89b0c2cccd97401bcf3f8b621e19e92