[roll] Roll fuchsia [inspect][drivers] Add an inspect-runner test that some drivers expose Inspect
This test is not very specific at all. It verifies that something in the
boot-drivers collection exposes something (but anything).

The idea is that if InspectSink is incorrectly routed to drivers, this
would hopefully fail.

Original-Fixed: 352803322
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185918
Original-Revision: ed3ac9bd70d7aa2a8c5230b1cd406f44d8f761c7
GitOrigin-RevId: cbbb6f61dbbac171c141ef84f4db2c3f4b37bd92
Change-Id: I3c6a138d19a52337f123e21939f62d7f16008da0