[roll] Roll fuchsia [starnix][cgroup] add_process takes TempRef<ThreadGroup> instead of WeakRef
We need TempRef to freeze the tasks of a newly added process if the
cgroup is frozen. We already upgrade the ThreadGroup to check that it is
alive, but previously we downgrade then upgrade again if we need to
freeze it. Pass in a TempRef directly to avoid the unnecessary
downgrade/upgrade.

Also remove freeze_process as it is no longer needed.

Original-Bug: 388081859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187812
Original-Revision: 623ae5ba9a854850a400cf6e9f40a35ddfd7c415
GitOrigin-RevId: e6cb718ce89bbb51a7ef38ed9079a0539ef0eb01
Change-Id: Ic04a0dbc4654a75b48ef92f6667f8ec507e3ed3b