[roll] Roll fuchsia [utest] Use relative offset for mapping
Restricted mode testing for aarch32 attempts to keep all memory shared
with the restricted environment under the addressable limit for that
code.  That limit is expressed as an absolute, but zx_vmar_map() expects
the offset to be relative to the base of the VMAR.

This error manifested as test flakiness because the memory was usually
mapped within the absolute range.  This change fixes the flakiness and
ensures any shared memory is mapped under the provided limit.

Original-Bug: 389697363
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187672
Original-Revision: 002a3db1992d68be99457ad5c5250f80153ec319
GitOrigin-RevId: 00199d100e858ce181e23a2d3aa2cf83300f8092
Change-Id: I6b44858186c827ce99ab081b6d558a9a39011d08