[roll] Roll fuchsia [power] Prepare the move of shutdown_watcher
This file is a copy from the src/power/power-manager/src. It is not
being used yet after this lands, until the next CL. In the next CL,
this copy of shutdown_watcher will be modified and the original will
be removed. In this way, gerrit will regard the file as a change
next time instead of as a new file.

Original-Bug: b/366035937
Test: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188193
Original-Revision: 1c346c8fc1309446331549cf77c6b4fcbe3d47b6
GitOrigin-RevId: 6c728e6b5dd257d76bd597c397a51febf657163b
Change-Id: I1beabea8fd29e12416dff03990ca1410e5b7fe41