[roll] Roll fuchsia [ffx] Making ffx_plugin deps explicit
The GN template ffx_plugin() adds 2 dependencies as part of the
implementation of the template. One of these ffx/core:lib is part
of the refactoring of the ffx framework.

This change removes the implicit adding of these dependencies and
makes it explicit in each BUILD.gn file for the subtools and plugins.

The advantage of this is that now changes can be made incrementally to
specific tools vs. having to change the template and related macros for
all tools at once.

This also cleans up some `use ffx_core as _;` statements where the
subtools no longer require the legacy ffx_plugin macros, which should
have a positive impact on the build graph.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188195
Original-Revision: 77d141651a89d9ab60dea42e5438ead1827a47d4
GitOrigin-RevId: f444471a98fe456959504d96ff9c1f37930c1c01
Change-Id: Ib5429148a756594765e42822c795ec5bd918fb01