[roll] Roll fuchsia [kernel][vm] Limit request dumping
Limits the amount of information emitted to the debuglog during a page
request overtime event by capping the number of items in lists that are
printed, and not redundantly dumping the source/provider state for
every waiting request.

Without this change, if there are multiple requests on a pager, then
each request prints out the information about every other request. This
causes the amount of information to be quadratic in the number of
requests, with most of it being duplicated and much of the remainder
being unhelpful.

Original-Bug: 381312579
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186933
Original-Revision: 9da14a8327aab820b16267f0fceee5f1b4f803f0
GitOrigin-RevId: 315a93a12af0b247d89ae66928fa0ca00b128b17
Change-Id: Ib29026d7eab18a56f5d1af0679b8c244210e79c2