[roll] Roll fuchsia [owners] Remove anmittal@google.com from src/lib/syslog/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179727
Original-Revision: 0933294c848c646ef77c3f2fcaa2d2225717b435
GitOrigin-RevId: c9c99fd66d1d8288dc6bfef416053cd322c6cd07
Change-Id: I0068f2aa2c03ab63442e133162f1a2a5326ba784