[roll] Roll fuchsia [ffx]: add structure to top level ffx errors
Introduces new type SerializableError and implements serde::Serialize for it.

Then the ffx_command::exit function now inspects the command line to see
if the machine writer was requested. If so it serializes the Error
variant to JSON and writes that to stderr. Otherwise, it just prints the
error string.

We need this additional type as we cannot directly serialize the
FfxError or ffx_command::Error since they have an `anyhow::Error`
embedded in them (which cannot be serialized).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1161889
Original-Revision: 4f5cedd88f89c7e3de35fa3c71a660994d7b33c0
GitOrigin-RevId: a928581dd9df8b753600da1bb4936349bc700693
Change-Id: I6a5444dc01b1b5adc026a31968870b7b063f4fc8