[roll] Roll fuchsia [netpol] Add CheckPresence to SocketProxy protocol
SocketProxy will be an optional component determined on a per-product
basis. To determine whether SocketProxy is available in the product, a
FIDL call must be made against the protocol to determine whether the
connection is valid.

The future client of FuchsiaNetworks, netcfg, uses this approach to
determine which filtering and DHCP protocol to use.

The existing endpoints offered by the API have side-effects (such as
writing to Inspect) so a new endpoint should exist for this behavior.

Original-Bug: 385368895
Test: Verified manually on https://fxrev.dev/1165212 with use of
CheckPresence

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186496
Original-Revision: f2ff3cd7a73ad38a5a138efca2ba24a108f5f26d
GitOrigin-RevId: 91a503471f32aa6c74b336f21bac0f9bd6a7f650
Change-Id: If228f325a36e4fab9e06b92f5e094588821b2273