[roll] Roll fuchsia [starnix] Refactor container namespace proxying
The ultimate goal of this change is to allow remotefs to take in
arbitrary paths, such as /userdata/data. If the root directory is a
valid namespace, then we'll create necessary subdirs (e.g. data) and
return them as the handle, per the existing contract.

Since we need 1..N access to arbitrary entries in the namespace, this
change refactors the whole of how the container passes along namespace
entries. Rather than transforming and relinquishing namespace handles
directly, it supplies the kernel with cloneable proxy references which
it can transform and dish out as needed.

In order to soft transition the existing remotefs creation (which only
mounts the entire source as `.`), I've placed logic within the remotefs
creation to assume the data directory. In a subsequent change, I'll be
migrating the fstab to specify the correct namespace and subdirs
explicitly, and then remove that logic.

Original-Bug: 379929394
Test: syscall mount_test, starnix-tests, manually verified boot process.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180339
Original-Revision: 851348cfb2e4d75f06cb231a6915ee5cfb706332
GitOrigin-RevId: fef7698f6f2908020f70b732f90524cd149e6d47
Change-Id: I35016e9f96fa1f24759a8b89cb047fc9376968e3