[roll] Roll fuchsia [ffx][ota] Allow specifying ota server port
This adds `--product-bundle-port`  option to `ffx target update` commands. This is the port
to start the OTA repo server on when using `--product_bundle`.

It can be configured be setting `repository.ota_port`. The default is 0,
which indicates a random unassigned port.

This is added to support using a specific port with other tools to
support remote workflows.

Original-Bug: 388844218
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190192
Original-Revision: 920268fa432de66f8f835256f435945d6c2eff94
GitOrigin-RevId: 2a1d211eae7ae139f8a35337e312e834d1bc356c
Change-Id: I51fa49921ee14c201969bfc53eae6fce138f1b8a