[roll] Roll fuchsia [kernel][vm] Make VmObject::user_id_ a relaxed atomic
Use a RelaxedAtomic to store the user_id_ instead of guarding with the
lock. The user_id_ is already const-ish and the lock is only preventing
read/write tearing with debug code, which can be achieved with an
atomic.

The parent_user_id_ for physical vmos is changed to be true const and
initialized in the constructor.

Not needing to acquire the lock of the parent when querying the
parent_user_id simplifies the locking semantics for fine grained VMO
locking.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187434
Original-Revision: 4e4d30dc80d82fae463b17534702f0ed827a10c6
GitOrigin-RevId: 4c38c1ec1bcfaa26447bba1648cd68f59fd75787
Change-Id: I7a416200f7b8ac1fc9a2100ff2890743e0ea4818