[roll] Roll fuchsia [owners] Remove pineapple@google.com from tools/shush/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189384
Original-Revision: b68487ef6bd349296390d4e76b946f448726fc80
GitOrigin-RevId: 323faa963f3d3acb79c1cb3906c9468d2f64c3a0
Change-Id: Id1296c1b2d97339449b01ce8eb0054849840e304