[roll] Roll fuchsia [owners] Remove pineapple@google.com from tools/rust_extract/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189383
Original-Revision: 38066fe2cbbb8ec341ffff6d2646bf47ef68f05f
GitOrigin-RevId: 2db70eabce0679ddb29d8cd4a16f9cf4ab009d20
Change-Id: Ibe950192b6b80efbd36c06b5fde921811e288d90