[roll] Roll fuchsia [display] Disallow existing IDs when importing events.
There was some confusion about the sematics of importing an event with
an existing ID.  This change aligns Coordinator.ImportEvent with the
rest of the Coordinator API, as well as Flatland.

All import APIs now return ZX_ERR_ALREADY_EXISTS when the ID is already
in use.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190198
Original-Revision: c339a199015d9014975e75f80872ccd2211fc883
GitOrigin-RevId: 43ef17e1ff17567629a6bb78035e28fed790d455
Change-Id: If2876bd9bb4e9f8e46b47ab2a3a330aa9752c886