[roll] Roll fuchsia [owners] Remove pineapple@google.com from scripts/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190593
Original-Revision: 0f382a3675c997a6481f553b1fceee8cebe155a6
GitOrigin-RevId: 1696fa4e7f6b5d7a358f2cf95774ca6420955dc2
Change-Id: I49896b17423ccd517a0358edd0977dab5a485343