[roll] Roll fuchsia [owners] Remove pineapple@google.com from build/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190592
Original-Revision: bcff5b907d0b680835cc5bb5ce92537587145eb5
GitOrigin-RevId: 9ceb74523d62a5d95da719ecf45b3d0e825dc928
Change-Id: Ia6985fa148c8be870c5f482d7b4ef0f570462e4b