[roll] Roll fuchsia [ffx][repo] Change default server mode to foreground
This changes the default server mode to be foreground. Previously it
was daemon mode  as the default, but now the migration is completed
to use foreground or background.

This is the next step in removing package serving from the daemon.

Original-Bug: 341095927
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185916
Original-Revision: 7e1a4fa55622c4ee38cb6b9229a0d3d245b5e6be
GitOrigin-RevId: 2490e0df7f093a51da2ece7531b40bb9bdac93e4
Change-Id: Iac1607f9b17202c4a2d464130f8f221d3e7f6a9f