[roll] Roll fuchsia [auth] Unset GOOGLE_APPLICATION_CREDENTIALS
Hide GOOGLE_APPLICATION_CREDENTIALS environment variable
from bazel when it is configured to use gcert-authentication.
This addresses the issue where bazel looks for this credential
file when it is not needed and fails ungracefully.

We still have to keep this variable for the exceptional
uses of remotetool, which doesn't have gcert-auth support
and needs the old gcloud method.
remotetool is used for on-demand downloading with
rbe_mode="low_bandwidth_remote".

Original-Bug: 358163278
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190203
Original-Revision: 2c4f91ff9a16e52b3bf4633e9909468a26c4f548
GitOrigin-RevId: 937bc841d90066172c7b66b5770c27597d806e06
Change-Id: I7cc35229f5ea02dfde620169b4b13647c1d7abca