[roll] Roll fuchsia [block_server][rust] Enable trace flows
Adds trace flow steps for the start and end of transactions, aligned
with the tracing available in the C++ block server.

I also changed the rust block client to attempt to reuse the request's
trace_flow_id so we can link the entire flow end-to-end.  Unfortunately,
a bit more work will be required here, since there's no way to specify
the trace flow ID to the block client.  I'll do that in a follow-up CL.
(For now, the behaviour is that the block client makes a new trace flow,
so there is a disjoint flow nested within the other flow, in processes
that implement block_server by forwarding to block_client (e.g. the GPT
component).

I also removed the future trace in
block_client::RemoteBlockClient::send, because it is redundant with the
flows.

Original-Bug: 375083001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191133
Original-Revision: 7d3d5c9cd843788b0ff6a92331560c570983ed17
GitOrigin-RevId: 91d65c8bbbb4132d4448b3c419f9f259bc714b5d
Change-Id: I7c265de9f91837078c5223d7e370e0a44a3b9b46