[roll] Roll fuchsia [driver_symbols] Remove asprintf from restricted symbols list
fuchsia-boot:///adc-buttons#meta/adc-buttons.cm was referencing this
symbol but it was in the restricted symbols list. It should be ok to
remove it from the list. It could be that this was transitively pulled
in via a libcxx change.

Original-Bug: 391390447
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191893
Original-Revision: 62d54a3900c1573388a462144572c6f0a337e639
GitOrigin-RevId: b57fcf6666049a94f0c3d4b44468a3452c9d902c
Change-Id: I2c9a8358f8da93ddfc4650948729c514c17395bc