[roll] Roll fuchsia [uart] Introduce KernelDriver/Driver::mmio_range
Introduce helper struct uart::MmioRange, which holds a tuple
reprensenting the memory range in the addressing provided by it's
configuration.

Introduce concept uart::MmioDriver<T> which is slightly
less verbose than `std::is_same_v<....., zbi_dcfg...>`
which evaluates the Driver::IoType(Pio, Mmio8, Mmio32) and decides
whether it is MMIO or not.

When a type complies with the MmioDriver concept, then uart::Driver and
uart::KernelDriver will provide additional accessors for obtaining the
device's MMIO range as 'uart::MMioRange mmio_range()'.

`io_slots` are adjusted to determine the proper length of the range.
The provided range is NOT required to be page aligned and it is
the callers responsibility to align to meet their requirements.

Original-Bug: 42084617
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121694
Original-Revision: 156cfd5e7b963ca4aa4e03e36328069fa6829e06
GitOrigin-RevId: 0286c5c6618346ee7b4d456797df3f2efa3bd735
Change-Id: I9587fbc9665f600d2d786bf24e2ff82ec3844009