[roll] Roll fuchsia [fxfs] clean up checksum chunking in multi_overwrite
The range math that broke up the vector of checksums that came back from
write was broken in a couple ways. First, it was calculating the device
range incorrectly because it was using a value that had an offset
included in it. Second, it wouldn't push the last chunk because it only
pushed when the bitmap value flipped.

This simplifies the logic so it's much more readable and understandable.
There are probably optimizations but it's not worth the obfuscation
here.

We also add a couple checks to the omni-test for multi_overwrite that
makes sure the checksums it puts into the transaction make sense to
prevent regressions here.

Original-Fixed: 383564671
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189239
Original-Revision: 5aec3d5b155136458b5b59210d189c1773e4fb12
GitOrigin-RevId: 1cf676c61811986f2039bce3f579fffbc0cdf3e2
Change-Id: I245cb81cf9f70f38ea06830ca926778e96b2be27