[roll] Roll fuchsia [shutdown-shim] Inspect for Collaborative Reboot
Integrate shutdown-shim with the Inspect framework, and add counters for
the currently scheduled collaborative reboots.

Because shutdown-shim is a built-in component of Component Manager, it
uses a local instance of the `Inspector` type, as opposed to the
globally-scoped `Inspector` (which is used by Component Manager). This
ensures that Component Manager and shutdown-shim each have their own
distinct Inspect sections.

Original-Bug: 380873116
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186272
Original-Revision: 1c90af568a88ac9620049f482154db74a8cc4674
GitOrigin-RevId: b48699985860c024d605e88da3d16d4d583db038
Change-Id: Ie37e30d06c908674cec5455498cb13fe72f46891