[roll] Roll fuchsia [bazel_sdk] avoid loading fuchsia_sdk in fuchsia_clang
This CL avoids loading the fuchsia_sdk repository inside of the
fuchsia_clang repository by hardcoding the list of supported CPUs. This
is safe because the CPUs will always exist in the official SDK. For
custom builds and in-tree builds where the CPUs might not exist it is
still safe as long as someone doesn't try to target the missing CPU.
This is guaranteed to be the case when building in-tree due to how the
build is structured and when using a custom CPU we document the
constraint.

Original-Bug: 369638604
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191925
Original-Revision: ad62d25b22b679df217c30de26f4cf5226c0daf5
GitOrigin-RevId: 12fbe58389a8e39a5cf8f1fdb8172a2c8126cf0c
Change-Id: Ic7e9c72e508ec3a66ad52bf9703b541353b9a68c