[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Use the NVIDIA device allocator for GPU malloc (#124277)
Summary:
This is a blocker on another patch in the OpenMP runtime. The problem is
that NVIDIA truly doesn't handle RPC-based allocations very well. It
cannot reliably update the MMU while a kernel is running and it will
usually deadlock if called from a separate thread due to internal use of
TLS.

This patch just removes the definition of `malloc` and `free` for NVPTX.
The result here is that they will be undefined, which is the cue for the
`nvlink` linker to define them for us. So, as far as `libc` is concerned
it still implements malloc.

GitOrigin-RevId: 9585321465c51900e9281edbc27dcd958a5ee973
Original-Revision: 15c0944acb56faa6b9409a9bafa997ff2caed7bd
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192486
Original-Revision: 9d8dd159a256b7c0938288c76f47d0eb2638fd01
Change-Id: I4ea02776494ddfc16dceefd621cb0b149885ce26