[roll] Roll fuchsia [virtual_audio] Remove thread sync for virtual audio drivers
Remove the thread synchronization mechanisms used on
`VirtualAudioDriver`. The virtual audio driver's dispatcher is
synchronous and can be treated as a single-threaded dispatcher and so
thread synchronization is not needed.

Make the implementation of `VirtualAudioDriver` decide which dispatcher
to use in order to perform the fuchsia.virtualaudio.Device getters and
setters.

Original-Bug: 388880875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191032
Original-Revision: 85d5353552d979e92961baa40e912ea111f706c3
GitOrigin-RevId: 584d4dd0fd6017e52396dfeb83fba102763810f8
Change-Id: Ie331f14a15b4c5197ea38227a2f1406b20696a58