[roll] Roll fuchsia [ffx][target_holders] Create TargetProxyHolder type
This is a newtype around TargetProxy which allows implementation of
TryFromEnv outside of the fho crate. This is the first "Holder" type, so
it has some duplication with the fho crate, but that will be
incrementally move out of fho in order to eliminate the dependencies
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Original-Bug: 391610009
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193235
Original-Revision: e338511ba819a86e604acfd47920023882355c72
GitOrigin-RevId: 1f7f15fe2ab05e580ed3f7830b2fe70cf6a7194e
Change-Id: I8d321556ceefd13d0eaa287d4b29118e2d3835fa