[roll] Roll fuchsia [drivers][power] Acquisition flips state flag
When we acquire a wake lease flip the flag tracking whether the
system is suspended. If the system was resumed when the
`fdf_power::WakeLease` was created, the object won't know the
system state until it changes. This might result in a lot of
unnecessary wake lease acquisitions.

MULTIPLY: aml-uart-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192485
Original-Revision: a92e98645dfa694160b9716ef2d84ff0c30a8344
GitOrigin-RevId: 1cf2352eea666582c14d76318ffdef79721701f0
Change-Id: Ic99614379867040c830f0a07dc11e0c28f10999e