[roll] Roll fuchsia [intel-display][hdmi] Read E-EDID when a display is available.
Previously, when an HDMI display was available, the intel-display
driver sent the DDC I2C interface to the display coordinator,
and the coordinator sent I2C commands to read the extended EDID
bytes.

With this change, HdmiDisplay reads the extended EDID bytes
when a display is available, and sends the bytes to the coordinator.

Test: core.x64 booted on NUC11TNHv5 with MIMO HDMI display.
Test: core.x64 booted on NUC11TNHv5 with HDMI headless plug.
Original-Bug: 390552176
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191929
Original-Revision: e482ba0dee068f8a8e41ad25ebcd9e493d9798ce
GitOrigin-RevId: 52fc376213647372a670c6a10cb28db3b5602024
Change-Id: I7ed9210cee52165691d91545b1f3c98ae954fb47