[roll] Roll fuchsia [ffx][target_holders] Create TargetInfoHolder type
This is a newtype around TargetInfo which allows implementation of
TryFromEnv outside of the fho crate.

This is part of eliminating the dependencies in fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Original-Bug: 392174744
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193395
Original-Revision: b982dd0531b3ce172e9c88a6c399ace3059d18ee
GitOrigin-RevId: 215e6a687bcf6dd6b95bb52e0d7ca50d974490f4
Change-Id: I1181c31e59a3887ccde5e1b1b4d519cdd36dd572