[roll] Roll fuchsia [ffx][target_holders] Create DaemonProxyHolder type
This is a newtype around DaemonProxy which allows implementation of
TryFromEnv outside of the fho crate.

This also refactors the daemon_protocol() method used to initialize
protocol proxies from the daemon connection.

This is part of eliminating the dependencies in fho which are related to
device interactions so they do not negatively impact build time of the
subtools that do not require devices.

Original-Bug: 392173931
Original-Bug: 392174459
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193587
Original-Revision: b73627f6580ffd7398171246c7c1f17a122798f8
GitOrigin-RevId: d4a38531b847020c9ed4b6a70709785f57f942a6
Change-Id: I7d8966587d43d606ac147834f39956af5edd132a