[roll] Roll fuchsia [ffx][fho] Refactoring toolbox() and moniker() out of fho
This refactors moniker() and toolbox() out of fho and into
target_holders, where it is primarily used when implementing
TryFromEnv for target related objects.

This is part of the project to eliminate the dependencies of fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Original-Bug: 392173451
Original-Bug: 392174212
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193571
Original-Revision: accf4488a16ea45a6b8ea13f2678aec2dc358de9
GitOrigin-RevId: 4a94af5e3f4b14956eb3beb01d61c272a978b8a8
Change-Id: I70679957204c9dcf8106859b28b82ae622e94479