[roll] Roll fuchsia [ffx][target_holders] Create RemoteControlProxyHolder type
This is a newtype around RemoteControlProxy which allows implementation of
TryFromEnv outside of the fho crate.

This is part of eliminating the dependencies in fho which are related to
device interactions so they do not negatively impact build time of the
subtools that do not require devices.

Original-Bug: 392174888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193755
Original-Revision: e15125a6d78b4153efc7847e01afc5921a72620d
GitOrigin-RevId: 1e0a65fd592181dfcfd0ee72340586eb0dddc8d1
Change-Id: I685f9db17f0aceb3e3377e476cb48869f0eb0b7c