[roll] Roll fuchsia [drivers] Use Open instead of Connect for dev-topological
As part of the ongoing fuchsia.io migration, there are new rights in
place for certain operations like getting attributes and enumerating
directory entries. This means that certain cases where opening a
directory as a service will fail, as we do not request any rights when
connecting to a service.

These rights are implicitly added in Open1 but are not added implicitly
in Open3. To avoid issues with switching fdio over to Open3, we ensure
that dev-topological is always opened as a directory with the correct
rights to directory watching can operate correctly.

In some cases it was easier to operate directly on the test realm's
exposed directory, so this was added where possible. In other cases,
we open the corresponding directory with Open3 instead of using Connect.
Also add watcher function for unowned client ends.

Original-Bug: 324111518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165732
Original-Revision: 67bc9ae70d755691dc763622e5e7174ea9e1e0f9
GitOrigin-RevId: 8bb52d83828259f5f5320a2da0b44e466d06c0c5
Change-Id: I84bf586521968586d6347159fb92a28a0603e416