[roll] Roll fuchsia [ffx] Fix expiration of "ad-hoc" targets
If you make an ad-hoc connection (e.g. "ffx -t 1.2.3.4 target echo"), we
will add 1.2.3.4 to the daemon's target collection, but due to incorrect
logic, it gets an "expiration", which means that if we don't see any
mDNS packets from that target (and there's no reason to expect any --
it wasn't "discovered"), then we eventually (after 25 seconds) expire
the address.

After which, the next request of "ffx -t 1.2.3.4 target echo" would
not find the target by its address, and it would establish a second
host-pipe connection. Then the daemon would wait for Overnet to
inform it of the new connection, so it could try to establish an RCS
connection. But from Overnet's point of view, there is no new target;
it already knew about it.

So the daemon never completes the RCS connection, resulting in a
time-out in the client.

The fix is simply to call it a "manual" target, with no expiry.

Original-Bug: 392682202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194236
Original-Revision: 74c1f638ff227ea78bcd886b6e72b1200d92c02f
GitOrigin-RevId: 6b10c7566e720db89856d9cef23ad59cafcb67e5
Change-Id: Iddc8fd73ae8deef788478ebebe9b86dec493dc03