[roll] Roll fuchsia [kernel] Move DumpRegistersAndBacktrace out of lockup_internal
DumpRegistersAndBacktrace is the platform and architecture independent
way to obtain a register dump and backtrace from a locked up CPU.
Move the function out of the lockup_internal namespace so that it's
clear it's intended to be used outside of the lockup detector.

Reorder declarations/definitions in diagnostics.h and diagnostics.cc.

No functional change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194247
Original-Revision: 3ef41712656fec503ed6c2682d7dc3693e0e46c4
GitOrigin-RevId: 6a9b9a3aad28e6acb4a9db1a22c625cbda508fb7
Change-Id: I9b0e6d78749db0e26092dee8bd7f3817c8fc24e1