[roll] Roll fuchsia [fdio] Don't use fdio_service_connect for opening /svc
The /svc should be opened as a directory and not a service. In most
cases, opening the /svc directory can be done with the component helper
library, or by using `fdio_service_connect` with the full path to the
service, including the /svc prefix.

Original-Bug: 324111518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165733
Original-Revision: 82f95b0273fca99f3366103c658c9f027bde7580
GitOrigin-RevId: 2138e1baa4f0f809e110cce5a90e5f109cc2e476
Change-Id: I1a2cb290a06713dc3224dd23e5984499faa8e571