[roll] Roll fuchsia [inspect] Improve Block API
These changes result in the removal of multiple runtime validations
which are lifted to be statically checked by the type system and as a
side effect the removal of various Results.

Improvements of 10% up to 70% in inspect writing benchmarks.

See:

- VIM3: https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket/8724426674705389009/+/u/compare_perf_test_results_without_and_with_CL/stdout

- X64: https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket/8724426674705389025/+/u/compare_perf_test_results_without_and_with_CL/stdout

Original-Bug: 392965471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193057
Original-Revision: 3bb766cc5be69f868ea37c4de073189fd6e2be0f
GitOrigin-RevId: 689867a7ac168b0a518566139fcd515190104174
Change-Id: Ic411e8071a470c4af8863abf112b697630978288