[roll] Roll fuchsia [intel-display][dp] Read E-EDID when a display is available.
Previously, when a DisplayPort display was available, the intel-display
driver sent the DDC I2C interface to the display coordinator,
and the coordinator sent I2C commands to read the extended EDID
bytes.

With this change, DpDisplay reads the extended EDID bytes when a
display is available, and sends the bytes to the coordinator.

Test: core.x64 booted on NUC11TNHv5 with Google P2718EC monitor
      (with E-EDID support).
Test: core.x64 booted on NUC11TNHv5 with HP ZR30w monitor (no
      E-EDID support).
Test: core.x64 booted on NUC11TNHv5 with DP headless plug (no
      E-EDID support).
Original-Bug: 390552176
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192957
Original-Revision: a708e382ea2da7c4808752f3e2e95c716bab8a13
GitOrigin-RevId: 3f168a3f986b4db72f6fc2db859c9250ea91b8f3
Change-Id: I27baa21834498989d334cde5e250ce7e3a3cbdec