[roll] Roll fuchsia [boot-shim] Introduce fallback for console arguments.
As a way to bridge the gap with linux bootloaders, we introduce a
fallback that interprets console=ttyXX arguments. For the time being we
take the index and assume it's the i-th non disabled uart in a
devicetree.

For x86 the index maps to the proper COM port, and the legacy
configurations are encoded in the table.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120687
Original-Revision: 388b15ff566005d740606b20d9c6c24de6932955
GitOrigin-RevId: 5273f6eebe8ce24ebd40e115fc1a9b6c44f75980
Change-Id: I5c6b04a9230e016ddb39aee725152885c3fc8599