[roll] Roll fuchsia [fxfs][fshost] Add support for per-container volumes
fshost now serves the fshost.StarnixVolumeProvider protocol, which
supports creating and/or mounting and unmounting a new volume. The
name of this starnix volume is defined by the product and plumbed
through to fshost via structured config. fshost no longer serves the
CryptManagement protocol.

The StarnixVolumeProvider currently only supports vending a single
volume, which is consumed by the main Starnix container. Expanding the
usage to support multiple volumes requires significant refactoring and
renaming of the StarnixVolume storage_config field.

The reason we are defining the Starnix volume name via a product config
as opposed to hardcoding a name in fshost is so that we can use a name
that we wouldn't have to change once we start supporting more volumes.

Original-Bug: b/389783709
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181143
Original-Revision: 70600dcc998bc55f142373e67abb06cc6b4f0bec
GitOrigin-RevId: 04538f1865a5b4122e8fb14e7987d59e41a16e2b
Change-Id: Ic433741e657917d903ed6d5a060a4b392202f576