[roll] Roll fuchsia [display][coordinator] Simplify integration tests opening sequence.
This CL removes the term "active client" from integration tests.  This
term was used to refer to two different concepts:

1. The Display Coordinator client that owns the displays has its
   configurations applied. Confusingly, the Coordinator's implementation
   uses `active_client_`, and the tests have been using
   IsActiveClient() for this concept.

2. VSync events are sent to the client that applied the displayed
   configuration. Comments have been using the term "active client"
   for this as well.

This CL replaces IsActiveClient(), which reached into the Display
Coordinator, with waiting for ownership notifications from the Display
Coordinator's FIDL API.

This CL modifies tests to apply a configuration (and wait for it to be
applied) when they expect VSync events. Waiting for a configuration to
be applied ensures the delivery of the SetVsyncEventDelivery() call
issued before applying the configuration.

Last, this CL removes the term "active" from comments.

Original-Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196697
Original-Revision: b519f23346fba44e8c85589459140e39606b1f7a
GitOrigin-RevId: 57b7c9a85bd1511b00de4390915fc7587cb23138
Change-Id: I30c3886a296fb09e33a5c34f85109479dff39951