[roll] Roll fuchsia [amlogic-display] Use monotonic timestamps for vsync interrupts
Modifies the `GetInterrupt` function to use the monotonic timeline
for timestamps. Usage of monotonic timestamps is required when
working with vsync timestamps (see attached bug for more info).

The RDMA and Display Capture interrupts, which also call
`GetInterrupt`, may be able to operate on the boot timeline, but
it is likely simpler to just make all interrupts operate on the
monotonic timeline.

Original-Bug: 385000927
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193399
Original-Revision: a0c046653682feb6d11050875930875df7d1c5b6
GitOrigin-RevId: e51c180ad7b49f800d17cd0ef469fa60199b8bee
Change-Id: I4d8a38937814af5463c5c57f697e0b8c71680342